Move helper methods to AbstractTestCase
